+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 ### Added
 - Added automated script for generating type stubs
 - Wrapped SCIPaddClique
+- Speed up MatrixExpr.sum(axis=...) via quicksum
 ### Fixed
 - all fundamental callbacks now raise an error if not implemented
+- Fixed the type of MatrixExpr.sum(axis=...) result from MatrixVariable to MatrixExpr.
+- Updated IIS result in PyiisfinderExec()
 ### Changed
 - changed default value of enablepricing flag to True
 ### Removed